sac works for basic envs

+++ torchlib/deep_rl/__init__.py
@@ -1,8 +1,9 @@
+from abc import ABC, abstractmethod
+
 import gym
 import numpy as np
 from gym.envs.registration import register
 from gym.spaces import Space
-from abc import ABC, abstractmethod
 from torchlib.utils.random import set_global_seeds
 
 register(
@@ -88,7 +89,7 @@ def test(env: gym.Env, agent: BaseAgent, num_episode=100, frame_history_len=1, r
         while not done:
             if render:
                 env.render()
-            action = agent.predict(np.concatenate(observation_lst, axis=-1).astype(np.float32))
+            action = agent.predict(np.concatenate(observation_lst, axis=-1))
             previous_observation, reward, done, _ = env.step(action)
             episode_reward += reward
             observation_lst.pop(0)
